ai_agent/requirements.txt
Flask==3.0.3
gunicorn==22.0.0
openai==1.35.3
pytest==8.3.1
flask-socketio==5.3.6
python-dotenv==1.0.1
sqlite3  # built-in
requests==2.32.3

